+++ src/main/kotlin/me/otomir23/sootychimneys/util/RandomOffset.kt
@@ -3,11 +3,11 @@ package me.otomir23.sootychimneys.util
 import net.minecraft.util.RandomSource
 
 object RandomOffset {
-    fun offset(input: Float, range: Float, random: RandomSource): Float {
-        return if (range <= 0.0) input else input + random.nextFloat(range * -1, range)
+    fun offset(input: Double, range: Double, random: RandomSource): Double {
+        return if (range <= 0.0) input else input + random.nextDouble(range * -1, range)
     }
 
-    fun RandomSource.nextFloat(min: Float, max: Float): Float {
-        return this.nextFloat() * (max - min) + min
+    fun RandomSource.nextDouble(min: Double, max: Double): Double {
+        return this.nextDouble() * (max - min) + min
     }
 }
